// src/@types/user.d.ts
export interface IPreRegister{
    email:string,
    password:string,
    confirmPassword:string,

}


export interface IRegister {
    firstName:string,
    lastName:string,
    email:string,
    phone:string,
    username:string,
    password:string,
    country:string
    company:string,
}


// export interface IAddUs {
//     firstName:string,
//     lastName:string,
//     email:string,
//     phone:string,
//     username:string,
//     password:string,
//     country:string
//     company:string,
// }